requirements.txt
# ─── Core Web Framework ───────────────────────────────────────────────────────
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9
aiofiles>=23.2.0
websockets>=12.0

# ─── Database ─────────────────────────────────────────────────────────────────
mysql-connector-python>=8.3.0

# ─── Queue / Cache ────────────────────────────────────────────────────────────
redis>=5.0.0

# ─── HTTP Client ──────────────────────────────────────────────────────────────
httpx>=0.26.0

# ─── NLP ──────────────────────────────────────────────────────────────────────
textblob>=0.18.0
nltk>=3.8.0

# ─── Config & Validation ──────────────────────────────────────────────────────
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# ─── Utilities ────────────────────────────────────────────────────────────────
python-dateutil>=2.8.0
rich>=13.7.0